app/requirements.txt
django==4.2.6
djangorestframework==3.14.0
djangorestframework-jsonapi==6.1.0
django-filter==23.3
environs==7.3.1
cryptography==41.0.3
